Break Monolith's realMain into an importable utility method for testing.
